+++ lib/calendar.rb
@@ -3,20 +3,41 @@ module ICalPal
   class Calendar
     include ICalPal
 
+    def [](k)
+      case k
+      when 'name', 'title'      # Aliases
+        @self['calendar']
+
+      else @self[k]
+      end
+    end
+
     QUERY = <<~SQL.freeze
 SELECT DISTINCT
 
-Store.name AS account,
-Calendar.title AS calendar,
-*
+s1.name AS account,
 
-FROM #{self.name.split('::').last}
+c1.UUID,
+c1.title AS calendar,
 
-JOIN Store ON store_id = Store.rowid
+c1.shared_owner_name,
+c1.shared_owner_address,
 
-WHERE Store.disabled IS NOT 1
-AND Store.display_order IS NOT -1
-AND Calendar.flags IS NOT 519
+c1.published_URL,
+c1.self_identity_email,
+c1.owner_identity_email,
+c1.notes,
+c1.subcal_account_id,
+c1.subcal_url,
+c1.locale
+
+FROM #{self.name.split('::').last} c1
+
+JOIN Store s1 ON c1.store_id = s1.rowid
+
+WHERE s1.disabled IS NOT 1
+AND s1.display_order IS NOT -1
+AND c1.flags IS NOT 519
 SQL
 
   end